# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GridScore AI API",
    "summary": "Credit infrastructure for Africa's energy lenders.",
    "description": "Cooperative PAYG repayment data platform. **All data in non-production environments is synthetic.**",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log in (OAuth2 password flow)",
        "operationId": "login_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_v1_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a refresh token",
        "operationId": "refresh_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/customers/{customer_id}": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Get a customer",
        "operationId": "get_customer_v1_customers__customer_id__get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/consents": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "List a customer's consent records",
        "operationId": "list_consents_v1_customers__customer_id__consents_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsentOut"
                  },
                  "title": "Response List Consents V1 Customers  Customer Id  Consents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customers"
        ],
        "summary": "Record a consent grant/denial",
        "operationId": "create_consent_v1_customers__customer_id__consents_post",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/customers/{customer_id}/scores": {
      "get": {
        "tags": [
          "customers"
        ],
        "summary": "Score history for a customer",
        "operationId": "score_history_v1_customers__customer_id__scores_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Customer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScoreHistoryItem"
                  },
                  "title": "Response Score History V1 Customers  Customer Id  Scores Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/events": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest repayment events inline (validated, anonymised, idempotent)",
        "operationId": "ingest_events_v1_ingest_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestEventsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batch": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Upload a CSV/JSON batch (processed by the background worker)",
        "operationId": "ingest_batch_upload_v1_ingest_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_batch_upload_v1_ingest_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IngestResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BatchAccepted"
                    }
                  ],
                  "title": "Response Ingest Batch Upload V1 Ingest Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/jobs/{job_id}": {
      "get": {
        "tags": [
          "ingestion"
        ],
        "summary": "Batch job status",
        "operationId": "job_status_v1_ingest_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "received": {
            "type": "integer",
            "title": "Received"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "received"
        ],
        "title": "BatchAccepted"
      },
      "Body_ingest_batch_upload_v1_ingest_batch_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingest_batch_upload_v1_ingest_batch_post"
      },
      "Body_login_v1_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_v1_auth_login_post"
      },
      "ConsentCreate": {
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/ConsentScope"
          },
          "granted": {
            "type": "boolean",
            "title": "Granted",
            "default": true
          },
          "source": {
            "type": "string",
            "maxLength": 80,
            "title": "Source",
            "default": "api"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "scope"
        ],
        "title": "ConsentCreate"
      },
      "ConsentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scope": {
            "$ref": "#/components/schemas/ConsentScope"
          },
          "granted": {
            "type": "boolean",
            "title": "Granted"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "granted",
          "source",
          "granted_at",
          "expires_at"
        ],
        "title": "ConsentOut"
      },
      "ConsentScope": {
        "type": "string",
        "enum": [
          "data_sharing",
          "enrichment",
          "scoring"
        ],
        "title": "ConsentScope",
        "description": "What a customer has consented to. Gates enrichment and scoring."
      },
...
      "CustomerOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "identity_hash": {
            "type": "string",
            "title": "Identity Hash"
          },
          "home_operator_id": {
            "type": "string",
            "format": "uuid",
            "title": "Home Operator Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "identity_hash",
          "home_operator_id",
          "created_at"
        ],
        "title": "CustomerOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestEventsRequest": {
        "properties": {
          "events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Events",
            "description": "Raw repayment rows to ingest"
          },
          "enrich": {
            "type": "boolean",
            "title": "Enrich",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "IngestEventsRequest"
      },
      "IngestResponse": {
        "properties": {
          "report": {
            "$ref": "#/components/schemas/IngestionReport"
          },
          "customers_enriched": {
            "type": "integer",
            "title": "Customers Enriched"
          },
          "signals_written": {
            "type": "integer",
            "title": "Signals Written"
          }
        },
        "type": "object",
        "required": [
          "report",
          "customers_enriched",
          "signals_written"
        ],
        "title": "IngestResponse"
      },
      "IngestionReport": {
        "properties": {
          "received": {
            "type": "integer",
            "title": "Received",
            "default": 0
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted",
            "default": 0
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates",
            "default": 0
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "default": 0
          },
          "customers_created": {
            "type": "integer",
            "title": "Customers Created",
            "default": 0
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/RowError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "title": "IngestionReport"
      },
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStatus"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "RiskTier": {
        "type": "string",
        "enum": [
          "A",
          "B",
          "C",
          "D",
          "E"
        ],
        "title": "RiskTier",
        "description": "Risk grade with documented PD cutoffs (defined in the scoring service)."
      },
      "RowError": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "index",
          "message"
        ],
        "title": "RowError"
      },
      "ScoreHistoryItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "view": {
            "$ref": "#/components/schemas/ScoreView"
          },
          "energy_credit_score": {
            "type": "integer",
            "title": "Energy Credit Score"
          },
          "default_probability": {
            "type": "number",
            "title": "Default Probability"
          },
          "risk_tier": {
            "$ref": "#/components/schemas/RiskTier"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "view",
          "energy_credit_score",
          "default_probability",
          "risk_tier",
          "model_version",
          "created_at"
        ],
        "title": "ScoreHistoryItem"
      },
...
      "ScoreView": {
        "type": "string",
        "enum": [
          "solo",
          "pooled"
        ],
        "title": "ScoreView",
        "description": "The two cooperative views \u2014 the core differentiator.\n\n``SOLO``   \u2014 features from the home operator's partial history only.\n``POOLED`` \u2014 features from the full cooperative history."
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "Token"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}